Feat(multiple): Implement DTOs for Accounts

+++ banco_cuenta/src/main/java/com/example/banco_cuenta/service/AccountService.java
@@ -1,7 +1,11 @@
 package com.example.banco_cuenta.service;
 
+import com.example.banco_cuenta.dto.AccountDTO;
+import com.example.banco_cuenta.dto.AccountDTOUpdate;
 import com.example.banco_cuenta.model.Account;
+import com.example.banco_cuenta.model.Bank;
 import com.example.banco_cuenta.repository.AccountRepository;
+import com.example.banco_cuenta.repository.BankRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +17,9 @@ public class AccountService {
     @Autowired
     private AccountRepository accountRepository;
 
+    @Autowired
+    private BankRepository bankRepository;
+
     public List<Account> findAll(){
         return accountRepository.findAll();
     }
@@ -21,11 +28,36 @@ public class AccountService {
         return accountRepository.findById(id);
     }
 
-    public Account save(Account account){
+    public Account save(AccountDTO accountDTO){
+        Account account = new Account();
+        account.setUserName(accountDTO.getUserName());
+        account.setUserLastname(accountDTO.getUserLastname());
+        account.setAccountNumber(accountDTO.getAccountNumber());
+        account.setPassword(accountDTO.getPassword());
+        account.setBalance(accountDTO.getBalance());
+        Optional<Bank> bank = bankRepository.findById(accountDTO.getBankId());
+        bank.ifPresent(account::setBank); //Try catch block missing to handle if is empty
         return accountRepository.save(account);
     }
 
-    public void deleteById(long id){
-        accountRepository.deleteById(id);
+    public Optional<Account> update(long id, AccountDTOUpdate accountDTOUpdate){
+        Optional<Account> account = accountRepository.findById(id);
+        if(account.isPresent()){
+            Account accountUpdate = account.get();
+            accountUpdate.setUserName(accountDTOUpdate.getUserName());
+            accountUpdate.setUserLastname(accountDTOUpdate.getUserLastname());
+            accountUpdate.setPassword(accountDTOUpdate.getPassword());
+            accountUpdate.setBalance(accountDTOUpdate.getBalance());
+            return Optional.of(accountRepository.save(accountUpdate));
+        }
+        return Optional.empty();
+    }
+
+    public boolean deleteById(long id){
+        if(findById(id).isPresent()){
+            accountRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
